Build a survival response from a clinical data frame

Build a survival response from a clinical data frame

Usage

survival_response(data, type = c("rfs", "os", "dfs"))

Arguments

data

Data frame such as the included clin_crc dataset.

type

Endpoint prefix: "rfs", "os", or "dfs". The columns <type>.delay and <type>.event must exist.

Value

A survival::Surv() object.

Examples

survival_response(clin_crc, type = "rfs")
#>    [1]   3.6400000   14.5300000+  16.4700000   19.7500000   20.0200000 
#>    [6]  23.9600000   27.0200000+  28.9600000   43.7900000   52.2400000 
#>   [11]  55.1300000   57.7900000   59.5300000   64.1000000   64.3700000 
#>   [16]  65.8800000   67.1900000   67.8200000   68.5100000   76.0700000 
#>   [21]  82.2900000   83.6000000   83.7300000   84.1300000  109.2100000 
#>   [26] 119.2100000  122.7200000  142.5500000           NA?  22.8000000 
#>   [31]  14.4000000           NA?  54.4000000   64.5000000    1.8000000 
#>   [36]  42.0000000   56.5000000   36.3000000           NA?  44.2000000 
#>   [41]  52.1000000   30.0000000   43.9000000   80.8000000    2.2600000 
#>   [46]   6.0100000+   9.9600000+  14.1300000+  14.2600000+  15.0500000 
#>   [51]  15.7100000   17.9500000   18.1800000   19.3600000+  22.0200000 
#>   [56]  22.3800000+  22.3800000+  24.1900000   25.2100000+  25.2400000 
#>   [61]  25.6100000   26.5300000   26.6600000   26.8200000   27.1500000 
#>   [66]  27.6400000+  28.6300000   28.8600000   29.2200000   31.0600000 
#>   [71]  33.2000000   33.6900000   35.9000000   36.7500000   37.3100000+
#>   [76]  37.3100000   39.2500000   44.9700000   46.7100000   47.8600000 
#>   [81]  49.8400000   52.1400000   52.5000000   55.3300000   55.9200000 
#>   [86]  58.0200000   59.6000000   59.9600000   63.2500000   64.8600000 
#>   [91]  70.6500000   76.0400000   79.1300000   80.2500000   85.2800000 
#>   [96]  86.4300000   88.9900000   93.6600000+ 106.9400000  110.7900000 
#>  [101] 118.5800000   12.1000000   29.6000000   37.0000000   28.5000000 
#>  [106]  56.3000000   47.4000000   13.5000000   56.8000000   52.8000000 
#>  [111]  45.3000000   55.7000000   40.0000000   14.2000000+   4.1000000 
#>  [116]  59.5000000   42.9000000   32.6000000+  31.3000000   33.9000000 
#>  [121]  47.7000000   38.2000000   44.4000000   44.8000000   57.0000000 
#>  [126]  40.4000000    5.2000000   11.6000000   58.4000000   36.2000000 
#>  [131]  45.4000000   50.9000000   55.9000000   14.3000000+  33.8000000 
#>  [136]  38.0000000   74.2000000   84.9000000    0.9200000    1.5700000+
#>  [141]   3.3200000+   4.2400000    5.2200000    5.3500000+   8.2100000+
#>  [146]   8.3100000+   8.5800000+   9.5600000+  11.9000000+  14.7200000+
#>  [151]  16.0700000+  18.9300000+  18.9600000   19.9800000   20.4400000 
#>  [156]  21.0400000+  21.2700000+  22.2500000   23.3400000+  24.1300000+
#>  [161]  24.8500000+  26.2600000   26.9200000   27.5800000+  27.7400000+
#>  [166]  29.3900000+  31.3600000   31.9200000   34.1900000+  36.9200000+
#>  [171]  36.9200000   38.0700000   38.7200000   41.0600000+  44.6700000 
#>  [176]  44.7400000   47.8000000   49.1100000   50.4300000   50.5900000 
#>  [181]  52.8600000   54.9000000   58.4500000   59.0700000   59.3400000 
#>  [186]  64.3300000   64.9300000   65.5500000   72.0900000   74.3600000 
#>  [191]  74.5300000   79.4600000+  85.6100000   85.8400000+  89.6200000 
#>  [196]  95.0700000   99.5100000  105.1700000  112.3300000   30.9000000 
#>  [201]  24.0000000   31.6000000   49.6000000   15.2000000+  24.2000000 
#>  [206]  69.1000000   68.1000000   10.3000000+  59.2000000   61.4000000 
#>  [211]  15.0000000   57.8000000   18.8000000    4.5000000+  38.1000000+
#>  [216]  40.4000000    5.7000000+  10.2000000   84.7000000   16.9000000 
#>  [221]   9.5000000+  68.7000000   59.5000000   18.7000000   70.8000000 
#>  [226]   4.4400000+  40.0000000+  72.9000000+  18.0000000+          NA?
#>  [231]          NA?          NA?          NA?          NA?          NA?
#>  [236]          NA?          NA?          NA?          NA?          NA?
#>  [241]          NA?          NA?          NA?          NA?          NA?
#>  [246]          NA?          NA?          NA?          NA?          NA?
#>  [251]          NA?          NA?          NA?          NA?          NA?
#>  [256]          NA?          NA?          NA?          NA?          NA?
#>  [261]          NA?          NA?          NA?          NA?          NA?
#>  [266]          NA?          NA?          NA?          NA?          NA?
#>  [271]          NA?          NA?          NA?          NA?          NA?
#>  [276]          NA?          NA?          NA?          NA?          NA?
#>  [281]          NA?          NA?          NA?          NA?          NA?
#>  [286]          NA?          NA?          NA?          NA?          NA?
#>  [291]  47.8000000+  48.4500000+  48.5300000+   5.2900000   28.9600000 
#>  [296]  41.0300000+  38.7300000+  36.2300000+  36.1000000+  29.8800000+
#>  [301]  38.3000000+  36.0300000+  38.3300000+  38.9200000+  40.1100000+
#>  [306]  36.3600000+  40.6000000   36.3600000+  36.1000000+  20.3800000 
#>  [311]  36.5900000+  11.6700000  112.1100000+  26.5600000+   6.4100000 
#>  [316] 109.8400000+ 100.8600000+  42.0500000   38.1400000+  99.1900000+
#>  [321]  33.0100000+ 101.8200000+   4.6000000   83.1100000+  96.8200000+
#>  [326]  96.8200000+  43.2700000+  66.7000000+   4.7000000+   2.1000000+
#>  [331]  27.7000000   39.2000000+  51.4000000+  53.8000000+  24.9000000+
#>  [336]  51.5000000+  39.2000000+  21.9000000   37.9000000+  16.6000000+
#>  [341]  49.0000000+  47.8000000+  93.7000000+  95.6000000+  59.1000000+
#>  [346]  64.3000000+  35.3000000   31.0000000+ 118.3000000+  44.3000000+
#>  [351]   6.8000000   11.9000000   33.1000000+  47.1000000+  71.9000000+
#>  [356]  65.9000000+  60.2000000+  14.9000000+  62.0000000+  75.8000000+
#>  [361]  65.3000000+  42.6000000    9.0000000  118.5000000+ 120.0000000+
#>  [366]  32.5000000+  49.1000000+  14.9000000    7.9000000+  11.1000000 
#>  [371]  14.6000000   14.6000000   33.5000000+  26.8000000+  80.5000000+
#>  [376]  39.1000000+  13.0000000  102.2000000+  15.8000000+  15.2000000+
#>  [381]  42.0000000+   1.8000000+  66.1000000+  55.6000000+  71.4000000+
#>  [386]  42.7000000+  13.6000000    1.9000000+  54.7000000+  32.4000000+
#>  [391]  39.6000000+  39.5000000+  59.5000000+  61.2000000+  61.5000000+
#>  [396]   9.1000000   17.9000000   53.6000000+  21.1000000   24.5000000+
#>  [401]  45.9000000+  52.7000000+  14.9000000   43.2000000+  29.8000000+
#>  [406]  40.6000000+   6.7000000   44.0000000+  11.5000000+  41.0000000+
#>  [411]  38.2000000+  34.7000000+  36.5000000+  35.4000000+   1.7000000?
#>  [416]  47.7000000+  47.6000000+  56.0547945+  39.8136986+   5.4246575 
#>  [421]   6.8712329   37.4794520   31.7260274   56.5808219   57.7643836+
#>  [426]   6.5424658    6.5753425   31.7260274   91.1671233+  41.3917808+
#>  [431]  85.4794521+  14.4328767   75.8794520+  11.7698630+  28.2082192 
#>  [436]  32.2849315+  41.3917808+  71.2767123+  63.8136986+  55.7260274+
#>  [441]  45.9945205+  15.1561644+  25.6109589   36.3945206+  36.1643836+
#>  [446]  35.6054794+   9.8958904   19.4301370    3.6493151   15.0904110+
#>  [451]  25.4794520+  53.2931507+  36.0328767+  40.7671233+  44.5479452+
#>  [456]  44.3178082+  45.0082192+  45.6986301+  43.5945205+  57.0739726+
#>  [461]  44.3506849+  42.2465753+  12.5917808   64.0438356+  55.8904110+
#>  [466]  51.0575343+  42.1479452+  33.0739726+  43.4958904+   8.6465753 
#>  [471]  47.8684931+  79.8575343+  87.2547945+  69.8301370+  45.6000000 
#>  [476]  47.4739726+  74.7616438+  61.7424658+  67.6931507+  15.4191781+
#>  [481]  29.5561644+  58.7835616+   0.5589041+  33.9287671+   1.2821918+
#>  [486]  26.4328767+  26.8273973+  19.3972603+  55.6602740+  60.2958904+
#>  [491]  50.0383562+  50.7945206+  69.0739726+  39.7808219+  50.4328767+
#>  [496]  12.0328767   47.0136986+  21.2054794   18.0821918    4.5041096 
#>  [501]  14.0054794   32.8438356+  55.3972603+  54.6739726+  55.2328767+
#>  [506]  63.7479452+  58.2246575+  58.4876712+  27.6493151   24.6904110 
#>  [511]  37.8082192   70.8493151+  62.5972603+  18.1150685   17.0630137 
#>  [516]  47.0794521+  14.4000000   59.8356164+  59.3095890+ 103.1342466+
#>  [521]  58.7506849+  54.8383562+  33.3698630   71.1123288+  60.9205479+
#>  [526]  36.0328767+  49.1835616+  92.6465753+  46.2575342+  36.1315069+
#>  [531]  94.1589041+  37.7753425+   6.7068493   38.0054795   43.6273973+
#>  [536]  32.6465753   46.5534247+  18.4438356   70.1917808+   5.2602740+
#>  [541]  11.0794521   19.9890411   62.0383562+  86.2684931+  94.0273973+
#>  [546]   2.5643836   11.5068493           NA+          NA+          NA+
#>  [551]          NA+          NA+          NA+          NA+          NA+
#>  [556]          NA+          NA+          NA+          NA           NA+
#>  [561]          NA           NA+          NA+          NA           NA 
#>  [566]          NA+          NA+          NA+          NA+          NA+
#>  [571]          NA           NA+          NA+          NA+          NA 
#>  [576]          NA+          NA           NA           NA+          NA+
#>  [581]          NA+          NA+          NA+          NA+          NA+
#>  [586]          NA+          NA+          NA+          NA+          NA 
#>  [591]          NA           NA           NA+          NA           NA+
#>  [596]          NA+          NA+          NA+          NA+          NA 
#>  [601]          NA?          NA?          NA?          NA?          NA?
#>  [606]          NA?          NA?          NA?          NA?          NA?
#>  [611]          NA?          NA?          NA?          NA?          NA?
#>  [616]          NA?          NA?          NA?          NA?          NA?
#>  [621]          NA?          NA?          NA?          NA?          NA?
#>  [626]          NA?          NA?          NA?          NA?          NA?
#>  [631]          NA?          NA?          NA?          NA?          NA?
#>  [636]          NA?          NA?          NA?          NA?          NA?
#>  [641]          NA?          NA?          NA?          NA?          NA?
#>  [646]          NA?          NA?          NA?          NA?          NA?
#>  [651]          NA?          NA?          NA?          NA?          NA?
#>  [656]          NA?          NA?          NA?          NA?          NA?
#>  [661]          NA?          NA?          NA?          NA?          NA?
#>  [666]          NA?          NA?          NA?          NA?          NA?
#>  [671]          NA?          NA?          NA?          NA?          NA?
#>  [676]          NA?          NA?          NA?          NA?          NA?
#>  [681]          NA?          NA?          NA?          NA?          NA?
#>  [686]          NA?          NA?          NA?          NA?          NA?
#>  [691]          NA?          NA?          NA?          NA?          NA?
#>  [696]          NA?          NA?          NA?          NA?          NA?
#>  [701]          NA?          NA?          NA?          NA?          NA?
#>  [706]          NA?          NA?          NA?          NA?          NA?
#>  [711]          NA?          NA?          NA?          NA?          NA?
#>  [716]          NA?          NA?          NA?          NA?          NA?
#>  [721]          NA?          NA?          NA?          NA?          NA?
#>  [726]          NA?          NA?          NA?          NA?          NA?
#>  [731]          NA?          NA?          NA?          NA?          NA?
#>  [736]          NA?          NA?          NA?          NA?          NA?
#>  [741]          NA?          NA?          NA?          NA?          NA?
#>  [746]          NA?          NA?          NA?          NA?          NA?
#>  [751]          NA?          NA?          NA?          NA?          NA?
#>  [756]          NA?          NA?          NA?          NA?          NA?
#>  [761]          NA?          NA?          NA?          NA?          NA?
#>  [766]          NA?          NA?          NA?          NA?          NA?
#>  [771]          NA?          NA?          NA?          NA?          NA?
#>  [776]          NA?          NA?          NA?          NA?          NA?
#>  [781]          NA?          NA?          NA?          NA?          NA?
#>  [786]          NA?          NA?          NA?          NA?          NA?
#>  [791]          NA?          NA?          NA?          NA?          NA?
#>  [796]          NA?          NA?          NA?          NA?          NA?
#>  [801]          NA?          NA?          NA?          NA?          NA?
#>  [806]          NA?          NA?          NA?          NA?          NA?
#>  [811]          NA?          NA?          NA?          NA?          NA?
#>  [816]          NA?          NA?          NA?          NA?          NA?
#>  [821]          NA?          NA?          NA?          NA?          NA?
#>  [826]          NA?          NA?          NA?          NA?          NA?
#>  [831]          NA?          NA?          NA?          NA?          NA?
#>  [836]          NA?          NA?          NA?          NA?          NA?
#>  [841]          NA?          NA?          NA?          NA?          NA?
#>  [846]          NA?          NA?          NA?          NA?          NA?
#>  [851]          NA?          NA?          NA?          NA?          NA?
#>  [856]          NA?          NA?          NA?          NA?          NA?
#>  [861]          NA?          NA?          NA?          NA?          NA?
#>  [866]          NA?          NA?          NA?          NA?          NA?
#>  [871]          NA?          NA?          NA?          NA?          NA?
#>  [876]          NA?          NA?          NA?          NA?          NA?
#>  [881]          NA?          NA?          NA?          NA?          NA?
#>  [886]          NA?          NA?          NA?          NA?          NA?
#>  [891]          NA?          NA?          NA?          NA?          NA?
#>  [896]          NA?          NA?          NA?          NA?          NA?
#>  [901]          NA?          NA?          NA?          NA?          NA?
#>  [906]          NA?          NA?          NA?          NA?          NA?
#>  [911]          NA?          NA?          NA?          NA?          NA?
#>  [916]          NA?          NA?          NA?          NA?          NA?
#>  [921]          NA?          NA?          NA?          NA?          NA?
#>  [926]          NA?          NA?          NA?          NA?          NA?
#>  [931]          NA?          NA?          NA?          NA?          NA?
#>  [936]          NA?          NA?          NA?          NA?          NA?
#>  [941]          NA?          NA?          NA?          NA?          NA?
#>  [946]          NA?          NA?          NA?          NA?          NA?
#>  [951]          NA?          NA?          NA?          NA?          NA?
#>  [956]          NA?          NA?          NA?          NA?          NA?
#>  [961]          NA?          NA?          NA?          NA?          NA?
#>  [966]          NA?          NA?          NA?          NA?          NA?
#>  [971]          NA?          NA?          NA?          NA?          NA?
#>  [976]          NA?          NA?          NA?          NA?          NA?
#>  [981]          NA?          NA?          NA?          NA?          NA?
#>  [986]          NA?          NA?          NA?          NA?          NA?
#>  [991]          NA?          NA?          NA?          NA?          NA?
#>  [996]          NA?          NA?          NA?          NA?          NA?
#> [1001]          NA?          NA?          NA?          NA?          NA?
#> [1006]          NA?          NA?          NA?          NA?          NA?
#> [1011]          NA?          NA?          NA?          NA?          NA?
#> [1016]          NA?          NA?          NA?          NA?          NA?
#> [1021]          NA?          NA?          NA?          NA?          NA?
#> [1026]          NA?          NA?          NA?          NA?          NA?
#> [1031]          NA?          NA?          NA?          NA?          NA?
#> [1036]          NA?          NA?          NA?          NA?          NA?
#> [1041]          NA?          NA?          NA?          NA?          NA?
#> [1046]          NA?          NA?          NA?          NA?          NA?
#> [1051]          NA?          NA?          NA?          NA?          NA?
#> [1056]          NA?          NA?          NA?          NA?          NA?
#> [1061]          NA?          NA?          NA?          NA?          NA?
#> [1066]          NA?          NA?          NA?          NA?          NA?
#> [1071]          NA?          NA?          NA?          NA?          NA?
#> [1076]          NA?          NA?          NA?          NA?          NA?
#> [1081]          NA?          NA?          NA?          NA?          NA?
#> [1086]          NA?          NA?          NA?          NA?          NA?
#> [1091]          NA?          NA?          NA?          NA?          NA?
#> [1096]          NA?          NA?          NA?          NA?          NA?
#> [1101]          NA?          NA?          NA?          NA?          NA?
#> [1106]          NA?          NA?          NA?          NA?          NA?
#> [1111]          NA?          NA?          NA?          NA?          NA?
#> [1116]          NA?          NA?          NA?          NA?          NA?
#> [1121]          NA?          NA?          NA?          NA?          NA?
#> [1126]          NA?          NA?          NA?          NA?          NA?
#> [1131]          NA?          NA?          NA?          NA?          NA?
#> [1136]          NA?          NA?          NA?          NA?          NA?
#> [1141]          NA?          NA?          NA?          NA?          NA?
#> [1146]          NA?          NA?          NA?          NA?          NA?
#> [1151]          NA?          NA?          NA?          NA?          NA?
#> [1156]          NA?          NA?          NA?          NA?          NA?
#> [1161]          NA?          NA?          NA?          NA?          NA?
#> [1166]          NA?          NA?          NA?          NA?          NA?
#> [1171]          NA?          NA?          NA?          NA?          NA?
#> [1176]          NA?          NA?          NA?          NA?          NA?
#> [1181]          NA?          NA?          NA?          NA?          NA?
#> [1186]          NA?          NA?          NA?          NA?          NA?
#> [1191]          NA?          NA?          NA?          NA?          NA?
#> [1196]          NA?          NA?          NA?          NA?          NA?
#> [1201]          NA?          NA?          NA?          NA?          NA?
#> [1206]          NA?          NA?          NA?          NA?          NA?
#> [1211]          NA?          NA?          NA?          NA?          NA?
#> [1216]          NA?          NA?          NA?          NA?          NA?
#> [1221]          NA?          NA?          NA?          NA?          NA?
#> [1226]          NA?          NA?          NA?          NA?          NA?
#> [1231]          NA?          NA?          NA?          NA?          NA?
#> [1236]          NA?          NA?          NA?          NA?          NA?
#> [1241]          NA?          NA?          NA?          NA?          NA?
#> [1246]          NA?          NA?          NA?          NA?          NA?
#> [1251]          NA?          NA?          NA?          NA?          NA?
#> [1256]          NA?          NA?          NA?          NA?          NA?
#> [1261]          NA?          NA?          NA?          NA?          NA?
#> [1266]          NA?          NA?          NA?          NA?          NA?
#> [1271]          NA?          NA?          NA?          NA?          NA?
#> [1276]          NA?          NA?          NA?          NA?          NA?
#> [1281]          NA?          NA?          NA?          NA?          NA?
#> [1286]          NA?          NA?          NA?          NA?          NA?
#> [1291]          NA?          NA?          NA?          NA?          NA?
#> [1296]          NA?          NA?          NA?          NA?          NA?
#> [1301]          NA?          NA?          NA?          NA?          NA?
#> [1306]          NA?          NA?          NA?          NA?          NA?
#> [1311]          NA?          NA?          NA?          NA?          NA?
#> [1316]          NA?          NA?          NA?          NA?          NA?
#> [1321]          NA?          NA?          NA?          NA?          NA?
#> [1326]          NA?          NA?          NA?          NA?          NA?
#> [1331]          NA?          NA?          NA?          NA?          NA?
#> [1336]          NA?          NA?          NA?          NA?          NA?
#> [1341]          NA?          NA?          NA?          NA?          NA?
#> [1346]          NA?          NA?          NA?          NA?          NA?
#> [1351]          NA?          NA?          NA?          NA?          NA?
#> [1356]          NA?          NA?          NA?          NA?          NA?
#> [1361]          NA?          NA?          NA?          NA?          NA?
#> [1366]          NA?          NA?          NA?          NA?          NA?
#> [1371]          NA?          NA?          NA?          NA?          NA?
#> [1376]          NA?          NA?          NA? 142.5500000+ 122.7200000+
#> [1381]  28.9600000+ 119.2100000+  59.5300000+  27.0200000   82.2900000+
#> [1386] 109.2100000+  65.8800000+  84.1300000+  16.4700000+  83.6000000+
#> [1391]  76.0700000+  67.8200000+  52.2400000+  64.1000000+  55.1300000+
#> [1396]  57.7900000+  50.1000000+  43.7900000+  19.7500000+   3.6400000+
#> [1401]  23.9600000+  20.0200000+ 118.5800000+  55.3300000+  93.6600000 
#> [1406] 110.7900000+  27.6400000  106.9400000+   0.0000000?  19.3600000 
#> [1411]  26.8200000+  52.5000000+   9.0500000   88.9900000+  64.8600000+
#> [1416]  14.1300000   85.2800000+  26.5300000+  37.3100000   86.4300000+
#> [1421]  55.9200000+  70.6500000+  76.0400000+  79.1300000+  22.3800000 
#> [1426]  80.2500000+  28.8600000+  25.2100000   31.0600000   37.3100000+
#> [1431]  22.0200000   63.2500000+  24.1900000+  52.1400000+  59.6000000+
#> [1436]  59.9600000+  47.8600000+  46.7100000+  45.9900000+  49.8400000+
#> [1441]  17.9500000+  15.7100000+  39.2500000+  36.7500000+  25.2400000+
#> [1446]  44.9700000+  35.9000000+  26.6600000+  29.2200000+  28.6300000+
#> [1451]   0.0000000?  33.2000000+  33.6900000+  25.6100000+  22.3800000 
#> [1456]  15.0500000+   2.2600000+  18.1800000+  27.1500000+  11.9000000 
#> [1461] 105.1700000+   8.2100000   29.3900000   95.0700000+ 112.3300000+
#> [1466]  21.2700000   16.0700000   27.7400000   14.7200000    5.3500000 
#> [1471]  99.5100000+  20.4400000+  27.5800000    9.5600000   23.3400000 
#> [1476]  89.6200000+  85.6100000+  64.3300000+  22.2500000+  74.5300000+
#> [1481]  59.0700000+  47.8000000+   3.3200000   72.0900000+  24.8500000 
#> [1486]  41.0600000   36.9200000   64.9300000+  50.5900000+   0.0000000?
#> [1491]  38.7200000+  54.9000000+  65.5500000+  58.4500000+  59.3400000+
#> [1496]  26.9200000+   8.5800000   26.2600000+  52.8600000+  21.0400000 
#> [1501]  44.6700000+  36.9200000+  50.4300000+   4.2400000+  49.1100000+
#> [1506]  38.0700000+  44.7400000+  34.1900000   24.1300000    0.9200000+
#> [1511]  31.3600000+  18.9600000+  85.8400000   31.9200000+  19.9800000+
#> [1516]   5.2200000+   0.0000000? 117.3000000+   0.0000000?   0.0000000?
#> [1521]   0.0000000?   0.0000000?   0.0000000?  82.6100000+   0.0000000?
#> [1526]   0.0000000?   0.0000000?   0.0000000?  10.6100000    0.0000000?
#> [1531]   0.0000000?   0.0000000?   0.0000000?  24.8800000   50.5300000+
#> [1536]   0.0000000?   0.0000000?   0.0000000?   0.0000000?   0.0000000?
#> [1541]  55.1900000+   0.0000000?   0.0000000?   0.0000000?   0.0000000?
#> [1546]   0.0000000?  13.0800000    0.0000000?   0.0000000?   9.9600000+
#> [1551]   8.9400000+  17.3200000    0.0000000?   0.0000000?   0.0000000?
#> [1556]  54.0493151+  63.9123288+  46.2904110+  60.0328767+  41.5890411 
#> [1561]  60.0328767+  60.2301370   57.0739726+  55.8575343+  64.8657534+
#> [1566]  54.7397260+  57.9616438+  57.2383562+  59.7698630+  62.4000000+
#> [1571]  54.5095890+   0.0000000    1.8082192+   2.9589041+   0.4931507+
#> [1576]  29.3917808   54.2794521+  34.9479452    0.0000000   46.0931507+
#> [1581]   0.0000000   45.4356164+   7.8246575+  63.9452055+  68.0547945+
#> [1586]  50.1369863+  18.0493151    0.0000000   46.4547945+  76.7671233+
#> [1591]   0.0000000   21.6328767+  33.1068493+   0.0000000   46.9808219+
#> [1596]   5.3260274+  49.7753425+  40.6684932+  40.7342466+   3.8794521 
#> [1601]  70.4547945+   0.0000000    0.0000000    0.0000000    0.0000000 
#> [1606]   0.0000000    0.0000000    0.0000000    0.4273973+  41.4246575+
#> [1611]   0.0000000    0.0000000    4.0000000   74.0000000+   1.0000000 
#> [1616]  73.0000000+  17.0000000   35.0000000+  33.0000000+  63.0000000+
#> [1621]  31.0000000?  93.0000000   63.0000000+  86.0000000+  65.0000000+
#> [1626]  75.0000000+  46.0000000+  56.0000000+  46.0000000+  36.0000000+
#> [1631]  49.0000000+ 109.0000000+  53.0000000+  59.0000000+  48.0000000+
#> [1636]  42.0000000+  87.0000000+   5.0000000    7.0000000+  68.0000000+
#> [1641]  68.0000000+  15.0000000   16.0000000   12.0000000+  55.0000000+
#> [1646]  20.0000000+  56.0000000+  43.0000000+  25.0000000+  76.0000000+
#> [1651]  89.0000000+  18.0000000    1.0000000+   9.0000000   73.0000000+
#> [1656]  25.0000000    0.0000000   89.0000000+   0.0000000    0.0000000 
#> [1661]   2.0000000+   2.0000000    0.0000000   10.0000000   18.0000000 
#> [1666] 101.0000000+  43.0000000+   0.0000000   91.0000000+  14.0000000 
#> [1671]  43.0000000+ 117.0000000+  21.0000000    9.0000000    5.0000000 
#> [1676]  82.0000000+   6.0000000   37.0000000   53.0000000+  32.0000000 
#> [1681]  94.0000000+  78.0000000+  36.0000000+  52.0000000+  55.0000000+
#> [1686]  60.0000000+  50.0000000+  50.0000000+  39.0000000+  50.0000000+
#> [1691]  33.0000000+   1.0000000+   0.0000000   63.0000000+  55.0000000+
#> [1696]  51.0000000+   0.0000000   42.0000000+   0.0000000   21.0000000 
#> [1701]  33.0000000+   0.0000000   26.0000000+  19.0000000+   0.0000000 
#> [1706]  15.0000000+  29.0000000+  26.0000000+  93.0000000+  92.0000000+
#> [1711]  26.0000000+  60.0000000+  59.0000000+  82.0000000+  43.0000000+
#> [1716]   0.0000000   70.0000000+   0.0000000   46.0000000+   0.0000000 
#> [1721] 183.0000000+   3.0000000  201.0000000+  68.0000000+ 183.0000000+
#> [1726] 172.0000000+ 163.0000000+   1.0000000   44.0000000+ 153.0000000+
#> [1731]  26.0000000   54.0000000  168.0000000+  53.0000000   26.0000000+
#> [1736] 119.0000000  127.0000000+ 129.0000000+  97.0000000+  43.0000000+
#> [1741]   8.0000000   79.0000000+  87.0000000+  69.0000000+  45.0000000 
#> [1746]  47.0000000+  74.0000000+  61.0000000+  67.0000000+   3.0000000 
#> [1751]  12.0000000   15.0000000+  15.0000000+  25.0000000+  21.0000000 
#> [1756]  46.0000000+  28.0000000   41.0000000+  32.0000000+   9.0000000 
#> [1761]  60.0000000+ 103.0000000+  36.0000000+   4.0000000   13.0000000 
#> [1766]  32.0000000+  18.0000000   46.0000000+  11.0000000    2.0000000 
#> [1771]  11.0000000   19.0000000   25.0000000   36.0000000+  35.0000000+
#> [1776]  19.0000000   71.0000000+  75.0000000+  14.0000000   63.0000000+
#> [1781]  85.0000000+  93.0000000+  11.0000000+  55.0000000+  86.0000000+
#> [1786]  61.0000000+  27.0000000    7.0000000    4.0000000   61.0000000 
#> [1791]   9.0000000   45.0000000   38.0000000   56.0000000   15.0000000 
#> [1796]   2.0000000  154.0000000+ 121.0000000+ 135.0000000+  66.0000000+
#> [1801] 140.0000000+ 127.0000000+ 132.0000000+ 109.0000000+  77.0000000+
#> [1806]  56.0000000+  66.0000000+   3.0000000    6.0000000  112.0000000+
#> [1811] 118.0000000+ 120.0000000+  93.0000000+  39.0000000    9.0000000 
#> [1816]  77.0000000+  41.0000000+   3.0000000   36.0000000    3.0000000 
#> [1821]  36.0000000    4.0000000  164.0000000+  73.0000000+ 130.0000000+
#> [1826]   4.0000000   15.0000000   97.0000000+  28.0000000   35.0000000 
#> [1831]  20.0000000   21.0000000   17.0000000   19.0000000   13.0000000 
#> [1836]  23.0000000   11.0000000    4.0000000   22.0000000   14.0000000 
#> [1841]   9.0000000   11.0000000  192.0000000+ 108.0000000+ 122.0000000+
#> [1846] 158.0000000+  73.0000000+  83.0000000+ 120.0000000+  73.0000000+
#> [1851]  50.0000000+  57.0000000+  66.0000000+  54.0000000+  67.0000000+
#> [1856]  53.0000000+  54.0000000+   4.0000000    7.0000000    0.0000000 
#> [1861]  34.0000000    0.0000000    0.0000000    0.0000000   57.0000000+
#> [1866]   0.0000000    0.0000000    0.0000000   47.0000000+   0.0000000 
#> [1871]   0.0000000    0.0000000    0.0000000    1.0000000+  65.0000000+
#> [1876]  57.0000000+  33.0000000+  20.0000000   37.0000000+  20.0000000+
#> [1881]  20.0000000+   1.0000000+   0.0000000+  27.0000000+  14.0000000+
#> [1886]  21.0000000+  15.0000000+  10.0000000+   7.0000000+  12.0000000+
#> [1891]  25.0000000+  19.0000000+  14.0000000+  14.0000000+  17.0000000+
#> [1896]  17.0000000+   2.0000000+  14.0000000+   0.0000000+   0.0000000+
#> [1901]   6.0000000+  14.0000000+  14.0000000+  12.0000000+   7.0000000+
#> [1906]  16.0000000+  16.0000000+   9.0000000    5.0000000+  13.0000000 
#> [1911]   9.0000000    0.0000000+  29.0000000+  13.0000000+   7.0000000+
#> [1916]   5.0000000+  76.0000000+  73.0000000+  60.0000000+  20.0000000+
#> [1921]  67.0000000+  48.0000000+  79.0000000+  65.0000000+  13.0000000 
#> [1926]  15.0000000+  16.0000000+  57.0000000+   3.0000000   43.0000000+
#> [1931]  51.0000000+  45.0000000+  52.0000000+  46.0000000+  56.0000000+
#> [1936]  51.0000000+  51.0000000+  27.0000000+   2.0000000+  37.0000000+
#> [1941]  43.0000000+  35.0000000+  17.0000000+  39.0000000+  10.0000000 
#> [1946]  25.0000000+  50.0000000+  25.0000000+  61.0000000+  22.0000000+
#> [1951]  46.0000000+   9.0000000+  41.0000000+  41.0000000+ 105.0000000+
#> [1956]  27.0000000+  19.0000000+  30.0000000+  31.0000000+  56.0000000+
#> [1961]  82.0000000+  92.0000000+  95.0000000+  25.0000000+ 106.0000000+
#> [1966]  97.0000000+  91.0000000+  46.0000000+  84.0000000+  83.0000000+
#> [1971]   5.0000000   36.0000000+  78.0000000+  10.0000000    9.0000000+
#> [1976]  66.0000000   80.0000000+  82.0000000+  53.0000000   85.0000000+
#> [1981]  83.0000000+  78.0000000   78.0000000+  86.0000000+  75.0000000+
#> [1986]  87.0000000+  80.0000000+  77.0000000+  46.0000000+  50.0000000+
#> [1991]  41.0000000+  74.0000000+  76.0000000+  48.0000000   69.0000000+
#> [1996]  28.0000000+  52.0000000+  75.0000000+  41.0000000   72.0000000+
#> [2001]   7.0000000+  62.0000000+   5.0000000+   9.0000000?  24.0000000+
#> [2006]  74.0000000  133.0000000+  23.0000000   17.0000000   53.0000000?
#> [2011]  42.0000000+  81.0000000+  12.0000000  141.0000000+  18.0000000 
#> [2016]  98.0000000+ 140.0000000+ 103.0000000+  47.0000000+  14.0000000 
#> [2021]  95.0000000+  92.0000000+  18.0000000   32.0000000   81.0000000+
#> [2026]  33.0000000   37.0000000   90.0000000+ 117.0000000+  87.0000000+
#> [2031]  82.0000000+  86.0000000+  95.0000000+   6.0000000   27.0000000 
#> [2036]  37.0000000   92.0000000+ 106.0000000+  63.0000000+  90.0000000+
#> [2041]  24.0000000   83.0000000+  93.0000000+   6.0000000   43.0000000 
#> [2046]  21.0000000    5.0000000    6.0000000  147.0000000+ 107.0000000+
#> [2051] 105.0000000+  64.0000000+  90.0000000+   8.0000000    0.0000000+
#> [2056]   0.0000000   56.0000000+  48.0000000+  79.0000000+ 106.0000000+
#> [2061] 132.0000000+  32.0000000+ 104.0000000+  91.0000000+ 125.0000000+
#> [2066]   5.0000000   16.0000000   53.0000000+   8.0000000    5.0000000 
#> [2071] 146.0000000+  70.0000000+  30.0000000    6.0000000   22.0000000 
#> [2076]  13.0000000   13.0000000  104.0000000+  59.0000000+  81.0000000+
#> [2081]  53.0000000+  55.0000000+  62.0000000+  53.0000000+  54.0000000+
#> [2086]  19.0000000    3.0000000   19.0000000   14.0000000    2.0000000 
#> [2091]   0.0000000    0.0000000    0.0000000   52.0000000+   0.0000000 
#> [2096]   0.0000000   48.0000000+   5.0000000   35.0000000+          NA?
#> [2101]  31.0000000+          NA?  19.0000000   19.0000000   97.0000000+
#> [2106]  99.0000000+ 100.0000000+  75.0000000+  74.0000000+   0.0000000+
#> [2111]  34.0000000   17.0000000+  12.0000000+  96.0000000+  96.0000000+
#> [2116]  88.0000000+          NA?  86.0000000+  84.0000000+  75.0000000+
#> [2121]  72.0000000+  24.0000000+  86.0000000+  85.0000000+          NA?
#> [2126]  80.0000000+  45.0000000    9.0000000   87.0000000+  64.0000000+
#> [2131]  81.0000000+  69.0000000+   3.0000000+  69.0000000+  65.0000000+
#> [2136]  42.0000000+  23.0000000   76.0000000+  75.0000000+  33.0000000+
#> [2141]  67.0000000+  49.0000000+  74.0000000+  74.0000000+  70.0000000+
#> [2146]  68.0000000+  59.0000000+  57.0000000+  18.0000000  142.0000000+
#> [2151]   1.0000000+  19.0000000   72.0000000+          NA? 106.0000000+
#> [2156] 131.0000000+ 146.0000000+  24.0000000  144.0000000+ 141.0000000+
#> [2161]  42.0000000+ 134.0000000+  14.0000000+  86.0000000+  47.0000000+
#> [2166]   4.0000000   13.0000000           NA? 118.0000000+  66.0000000+
#> [2171]  32.0000000   15.0000000   11.0000000   62.0000000+  76.0000000+
#> [2176]  99.0000000+   0.0000000    4.2000000   75.5000000+  98.2000000+
#> [2181]  67.3000000    1.0000000   74.1000000+  17.5000000   29.7000000 
#> [2186]   0.0000000   72.5000000+  36.0000000+  34.2000000+  64.8000000+
#> [2191]  31.6000000?  96.8000000+  95.3000000   64.8000000+  88.0000000+
#> [2196]  66.7000000+  76.7000000+  47.1000000+  57.3000000+  47.7000000+
#> [2201]  37.1000000+  50.3000000+ 111.0000000+  54.1000000+  59.9000000+
#> [2206]  49.1000000+  43.0000000+  88.5000000+   5.1000000    7.9000000+
#> [2211]  69.9000000+  69.9000000+  15.9000000   16.6000000   12.9000000+
#> [2216]  56.5000000+  10.7000000   21.3000000+  57.2000000+  44.0000000+
#> [2221]  99.0000000+  25.9000000+  77.5000000+  90.4000000+  18.6000000 
#> [2226]  11.0000000    1.6000000+   9.5000000   74.2000000+  25.8000000 
#> [2231]   0.0000000   90.3000000+   0.0000000    0.0000000    2.1000000+
#> [2236]   2.0000000    8.6000000   10.3000000   18.3000000  103.0000000+
#> [2241]  44.0000000+   0.0000000   93.1000000+  14.8000000           NA?
#> [2246]  17.8333333   16.7666667+  17.4333333+  28.2000000+  14.9000000+
#> [2251]  18.0333333+  14.5666667+  19.8333333+  17.4000000+  18.4000000+
#> [2256]  12.7000000   10.0000000   12.9000000+  15.4666667+  11.1000000+
#> [2261]   9.3000000+  10.4666667+   8.2333333+  12.1333333    7.7000000+
#> [2266]  15.4000000+   9.5333333+  11.4000000+  10.2000000+  15.1000000+
#> [2271]   6.1000000   13.4666667+  12.0666667+  10.6666667+  11.6000000+
#> [2276]   9.3333333+  11.8666667+   8.5000000+  12.2666667+  11.4666667+
#> [2281]  10.9333333+   9.2666667+   8.6333333+  15.2000000+   9.4000000+
#> [2286]  13.3000000+  11.8000000+   9.7666667+   6.3000000+  14.6333333 
#> [2291]   8.5666667   35.6333333   55.9333333    9.7666667+          NA?
#> [2296]   7.1333333?   3.0666667+          NA?  37.5666667+          NA?
#> [2301]  35.5000000?  58.8333333?  63.8333333+   1.9666667    1.0333333?
#> [2306]  13.2000000?  27.3666667           NA?   9.2000000+  24.3666667?
#> [2311]          NA?  37.5666667+  36.5333333+   8.1666667+          NA?
#> [2316]          NA?  19.3333333+  34.5000000+  29.4000000?  29.4000000+
#> [2321]  26.3666667+  13.1666667+   1.0000000?  14.2000000+  34.4666667+
#> [2326]  21.3000000+   4.0666667   24.3333333+  18.2000000+   3.0333333 
#> [2331]  23.3333333?  20.2666667+  14.1333333+  20.2666667?  61.8666667+
#> [2336]  79.1666667+   6.1333333+   7.0666667+  54.7666667?          NA?
#> [2341]  14.2000000+   0.9333333?  50.7333333+  47.6666667+  15.2333333+
#> [2346]  11.1666667+   6.0666667+          NA?  19.2666667+  13.2000000 
#> [2351]          NA?          NA?   5.1000000?  86.2333333?  27.3666667+
#> [2356]          NA?  19.3000000+   5.1000000   10.2000000?  25.4000000 
#> [2361]          NA?  33.5000000+          NA?  25.3666667+   1.0333333?
#> [2366]  18.2333333+  16.1666667+  39.5333333+  37.4666667+  13.2000000 
#> [2371]  12.1666667           NA?  17.2666667?  10.2000000?          NA?
#> [2376]  33.5333333?          NA?  36.5333333+  32.5000000?          NA?
#> [2381]  31.5000000+  31.5000000+  30.4666667+  30.4666667+  53.7333333+
#> [2386]  17.2666667+  13.1333333   27.4000000+  30.4000000+          NA?
#> [2391]  18.3000000?  31.4333333+   2.0333333+  13.1666667?  24.3333333?
#> [2396]  33.4666667+  26.3666667+  24.3333333+          NA?  21.2666667+
#> [2401]  34.5000000?          NA?  22.3000000+  36.5333333+  16.3000000+
#> [2406]  40.5333333   13.2333333+  34.5333333+  26.3666667+  25.3666667+
#> [2411]  24.3333333+  27.4000000+  27.4000000?          NA?  19.3333333+
#> [2416]   8.0666667?  27.4000000+  14.1333333+  38.5666667?  19.2666667+
#> [2421]  30.4333333+  34.5000000+  18.3000000+  18.3000000?  38.5666667+
#> [2426]          NA?  27.4000000+  42.6000000+          NA?  17.2666667+
#> [2431]  15.2000000+  22.3000000+  35.4666667+  15.2333333+  15.2333333+
#> [2436]   6.1000000?  32.4666667?  12.1666667+  31.4333333+  31.4333333+
#> [2441]  25.4000000           NA?  20.2666667           NA?  33.5000000+
#> [2446]          NA?  26.3666667   37.5333333           NA?  39.6000000+
#> [2451]  52.7000000+          NA?  40.5333333+          NA?          NA?
#> [2456]  10.2000000           NA?          NA?  33.5000000   18.3000000?
#> [2461]  41.5333333+  12.3333333+  21.6666667+  11.4333333   50.5333333 
#> [2466]  24.8666667+  25.4333333+   5.8666667+  15.6000000   27.8000000+
#> [2471]  10.8333333   21.6000000   16.6666667+          NA?          NA?
#> [2476]          NA?          NA?          NA?          NA?          NA?
#> [2481]          NA?          NA?          NA?          NA?          NA?
#> [2486]          NA?          NA?          NA?          NA?          NA?
#> [2491]          NA?          NA?          NA?          NA?          NA?
#> [2496]          NA?          NA?          NA?          NA?          NA?
#> [2501]          NA?          NA?          NA?          NA?          NA?
#> [2506]          NA?          NA?          NA?          NA?          NA?
#> [2511]          NA?          NA?          NA?          NA?          NA?
#> [2516]          NA?          NA?          NA?          NA?          NA?
#> [2521]          NA?          NA?          NA?          NA?          NA?
#> [2526]          NA?          NA?          NA?          NA?          NA?
#> [2531]          NA?          NA?          NA?          NA?          NA?
#> [2536]          NA?          NA?          NA?          NA?          NA?
#> [2541]          NA?          NA?          NA?          NA?          NA?
#> [2546]          NA?          NA?          NA?          NA?          NA?
#> [2551]          NA?          NA?          NA?          NA?          NA?
#> [2556]          NA?          NA?          NA?          NA?          NA?
#> [2561]          NA?          NA?          NA?          NA?          NA?
#> [2566]          NA?          NA?          NA?          NA?          NA?
#> [2571]          NA?  27.4666667?   6.2000000    0.9666667+  15.9666667+
#> [2576]   0.2000000?  14.1000000+  12.0000000+  13.2000000   68.2666667+
#> [2581]  51.0666667+  59.2000000+   1.4333333+   5.7333333+  22.9333333+
#> [2586]   5.2000000+          NA?  13.7000000   10.1333333   62.2666667+
#> [2591]  50.1000000+   6.8666667+          NA+  75.6666667   29.9666667+
#> [2596]   5.3000000+  11.9000000+   3.2333333+   1.9666667+          NA?
#> [2601]   2.2666667+   3.1000000+   2.1666667+   3.3000000+   3.3333333+
#> [2606]  12.7666667+  12.3666667+  12.9333333+   6.9000000   11.8000000 
#> [2611]          NA?          NA?          NA?          NA?          NA?
#> [2616]          NA?          NA?          NA?          NA+  16.0000000 
#> [2621]          NA+  49.7666667+  52.0333333+  10.1333333?   0.0000000?
#> [2626]   9.7333333           NA+          NA+ 126.0000000+          NA?
#> [2631]          NA?          NA?   6.1333333+  23.3000000+   8.1333333+
#> [2636]   7.1333333+   8.1333333+  20.2333333+   4.1000000+   9.1000000+
#> [2641]  10.1333333+  13.1000000+  11.1666667+  10.1333333+  11.1666667 
#> [2646]   1.0333333   10.1333333+   2.0333333+  14.2000000+   7.0666667+
#> [2651]   7.0666667+  14.2333333+  16.2333333+   7.0666667+  14.1333333+
#> [2656]   4.0666667+   6.0333333+   8.1000000+   8.0666667+   6.0666667+
#> [2661]   7.0666667+   5.0000000+   5.1000000+   5.1000000+   5.1000000+
#> [2666]   4.0666667+   4.0666667+   3.0333333+  24.0666667   12.4333333+
#> [2671]  19.3000000   13.7000000+  11.6666667+  12.8666667   20.7000000+
#> [2676]  18.0000000+  18.5000000+  25.8333333+  43.8666667+  15.3333333+
#> [2681]  16.4333333    4.1000000   17.3666667+  12.6666667+  16.3666667+
#> [2686]  15.8000000+   7.6333333+  12.5666667+  10.2666667+  14.5000000+
#> [2691]   9.1666667+   9.5666667+  11.8000000+  13.6000000+  13.5333333+
#> [2696]  12.1666667+  11.5333333+          NA?          NA?          NA?
#> [2701]          NA?          NA?          NA?          NA?          NA?
#> [2706]          NA?          NA?          NA? 121.3666667+  19.1333333 
#> [2711]  38.7333333+ 132.4666667+  64.0000000   50.6000000+   2.6666667 
#> [2716]  12.7666667+ 142.3333333+   7.6000000+  44.9333333+ 133.3333333+
#> [2721] 137.5333333+   7.8666667+ 141.1000000+   5.9666667+  10.6666667 
#> [2726]  61.6333333+  16.1333333   96.5000000+          NA?          NA?
#> [2731]          NA?          NA?          NA?          NA?          NA?
#> [2736]          NA?          NA?          NA?          NA?          NA?
#> [2741]          NA?          NA?          NA?          NA?          NA?
#> [2746]          NA?          NA?          NA?          NA?          NA?
#> [2751]          NA?          NA?          NA?   8.7333333+  10.4333333 
#> [2756]   9.6000000   13.8333333+  13.6333333+   6.2666667+  36.3666667+
#> [2761]          NA?  16.5333333+  11.2000000   16.4666667+  34.1333333+
#> [2766]  13.4333333+   0.5333333+  35.9000000+  19.7000000+          NA?
#> [2771]          NA?          NA?          NA?          NA?          NA?
#> [2776]          NA?          NA?          NA?          NA?          NA?
#> [2781]          NA? 109.1666667+  17.2333333    8.4666667+  14.6333333 
#> [2786]   9.7000000   75.6000000+  68.2333333+  29.0666667   28.6333333 
#> [2791]  45.3000000+  55.8000000+  52.4666667+  45.7333333+  39.9666667+
#> [2796]   1.2666667   37.2333333+  19.7333333   26.8000000+  22.4000000+
#> [2801]  26.3000000   13.9666667+  36.3000000+  26.5333333+  85.4666667 
#> [2806]  47.4000000+  33.4333333   53.1666667+          NA?          NA?
#> [2811]          NA?  18.4333333+  19.6000000+  43.3666667+  33.4333333+
#> [2816]  31.7333333+